requirements/dev.txt
-r base.txt

pytest==7.2.0
flake8==6.0.0
mypy==1.1.1
black[jupyter]==22.12.0
types-PyYAML==6.0.12.2
assertpy==1.1
model-bakery==1.10.1
sqlalchemy2-stubs
